rename star and vantage _backend
