+++ pylabrobot/liquid_handling/backends/__init__.py
@@ -1,8 +1,8 @@
 from .backend import LiquidHandlerBackend
 from .chatterbox import LiquidHandlerChatterboxBackend
 from .chatterbox_backend import ChatterBoxBackend
-from .hamilton.STAR import STAR
-from .hamilton.vantage import Vantage
+from .hamilton.STAR_backend import STAR
+from .hamilton.vantage_backend import Vantage
 from .http import HTTPBackend
 from .opentrons_backend import OpentronsBackend
 from .saver_backend import SaverBackend
