+++ pylabrobot/liquid_handling/backends/hamilton/__init__.py
@@ -2,5 +2,5 @@
 
 from .base import HamiltonLiquidHandler
 from .pump import Pump  # TODO: move elsewhere.
-from .STAR import STAR
-from .vantage import Vantage
+from .STAR_backend import STAR
+from .vantage_backend import Vantage
